+++ R/DistToNearest.R
@@ -641,15 +641,15 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 #'
 #' Get non-zero distance of every heavy chain (\code{IGH}) sequence (as defined by 
 #' \code{sequenceColumn}) to its nearest sequence in a partition of heavy chains sharing the same 
-#' V gene, J gene, and junction length (VJL), or in a partition of single cells with heavy chains
-#' sharing the same heavy chain VJL combination, or of single cells with heavy and light chains 
-#' sharing the same heavy chain VJL and light chain VJL combinations.
+#' V gene, J gene, and junction length (V-J-length), or in a partition of single cells with heavy/long chains
+#' sharing the same heavy/long chain V-J-length combination, or of single cells with heavy/long and light/short chains 
+#' sharing the same heavy/long chain V-J-length and light/short chain V-J-length combinations.
 #'
 #' @param    db              data.frame containing sequence data.
 #' @param    sequenceColumn  name of the column containing the junction for grouping and for calculating
-#'                           nearest neighbot distances. Note that while both heavy and light chain junctions
-#'                           may be used for VJL grouping, only the heavy chain junction is used to calculate 
-#'                           distances.
+#'                           nearest neighbor distances. Note that while both heavy/long and light/short chain junctions
+#'                           may be used for V-J-length grouping, only the heavy/long chain (IGH, TRB, TRD) junction is 
+#'                           used to calculate distances.
 #' @param    vCallColumn     name of the column containing the V-segment allele calls.
 #' @param    jCallColumn     name of the column containing the J-segment allele calls.
 #' @param    model           underlying SHM model, which must be one of 
@@ -664,11 +664,11 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 #'                           is used. if \code{FALSE} the union of ambiguous gene 
 #'                           assignments is used to group all sequences with any 
 #'                           overlapping gene calls.
-#' @param    VJthenLen       a Boolean value specifying whether to perform partitioning as a 2-stage
+#' @param    VJthenLen       logical value specifying whether to perform partitioning as a 2-stage
 #'                           process. If \code{TRUE}, partitions are made first based on V and J
-#'                           annotations, and then further split based on junction lengths corresponding 
+#'                           gene, and then further split based on junction lengths corresponding 
 #'                           to \code{sequenceColumn}. If \code{FALSE}, perform partition as a 1-stage 
-#'                           process during which V annotation, J annotation, and junction length are used 
+#'                           process during which V gene, J gene, and junction length are used 
 #'                           to create partitions simultaneously. Defaults to \code{TRUE}.
 #' @param    nproc           number of cores to distribute the function over.
 #' @param    fields          additional fields to use for grouping.
@@ -677,57 +677,65 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 #' @param    mst             if \code{TRUE}, return comma-separated branch lengths from minimum 
 #'                           spanning tree.
 #' @param    subsample       number of sequences to subsample for speeding up pairwise-distance-matrix calculation. 
-#'                           Subsampling is performed without replacement in each VJL group of heavy chain sequences. 
+#'                           Subsampling is performed without replacement in each V-J-length group of heavy chain sequences. 
 #'                           If \code{subsample} is larger than the unique number of heavy chain sequences in each 
 #'                           VJL group, then the subsampling process is ignored for that group. For each heavy chain
 #'                           sequence in \code{db}, the reported \code{dist_nearest} is the distance to the closest
-#'                           heavy chain sequence in the subsampled set for the VJL group. If \code{NULL} no 
+#'                           heavy chain sequence in the subsampled set for the V-J-length group. If \code{NULL} no 
 #'                           subsampling is performed.
 #' @param    progress        if \code{TRUE} print a progress bar.
-#' @param    cellIdColumn    name of the character column containing cell IDs. Only applicable and required for 
-#'                           single-cell mode.
-#' @param    locusColumn     name of the character column containing locus information. Only applicable and 
-#'                           required for single-cell mode.
-#' @param    groupUsingOnlyIGH    use only heavy chain (\code{IGH}) sequences for VJL grouping, disregarding 
-#'                                light chains. Only applicable and required for single-cell mode. 
-#'                                Default is \code{TRUE}. Also see \link[alakazam]{groupGenes}.                   
-#' @param    keepVJLgroup         a Boolean value specifying whether to keep in the output the the column 
-#'                                column indicating grouping based on VJL combinations. Only applicable for
-#'                                1-stage partitioning (i.e. \code{VJthenLen=FALSE}). Also see 
-#'                                \link[alakazam]{groupGenes}.
+#' @param    cellIdColumn    name of the character column containing cell identifiers or barcodes. 
+#'                           If specified, grouping will be performed in single-cell mode
+#'                           with the behavior governed by the \code{locusColumn} and 
+#'                           \code{onlyHeavy} arguments. If set to \code{NULL} then the 
+#'                           bulk sequencing data is assumed.
+#' @param    locusColumn     name of the column containing locus information. 
+#'                           Only applicable to single-cell data.
+#'                           Ignored if \code{cellIdColumn=NULL}.
+#' @param    onlyHeavy       use only the IGH (BCR) or TRB/TRD (TCR) sequences 
+#'                           for grouping. Only applicable to single-cell data.
+#'                           Ignored if \code{cellIdColumn=NULL}. 
+#'                           See \link[alakazam]{groupGenes} for further details.               
+#' @param    keepVJLgroup    logical value specifying whether to keep in the output the the column 
+#'                           column indicating grouping based on V-J-length combinations. Only applicable for
+#'                           1-stage partitioning (i.e. \code{VJthenLen=FALSE}). Also see 
+#'                           \link[alakazam]{groupGenes}.
 #' 
 #' @return   Returns a modified \code{db} data.frame with nearest neighbor distances between heavy chain
 #'           sequences in the \code{dist_nearest} column if \code{cross=NULL}. If \code{cross} was 
 #'           specified, distances will be added as the \code{cross_dist_nearest} column. 
 #'           
-#'           Note that distances between light chain sequences are not calculated, even if light chains 
-#'           were used for VJL grouping via \code{groupUsingOnlyIGH=FALSE}. Light chain sequences, if any,
-#'           will have \code{NA} in the \code{dist_nearest} field.
+#'           Note that distances between light/short (IGK, IGL, TRA, TRG) chain sequences are not calculated, 
+#'           even if light/short chains were used for V-J-length grouping via \code{onlyHeavy=FALSE}. 
+#'           Light/short chain sequences, if any, will have \code{NA} in the \code{dist_nearest} output column.
 #'           
 #'           Note that the output \code{vCallColumn} and \code{jCallColumn} columns will be converted to 
-#'           \code{character} if they were \code{factor} in the input \code{db}.
+#'           type \code{character} if they were type \code{factor} in the input \code{db}.
 #'
 #' @details
+#' To invoke single-cell mode the \code{cellIdColumn} argument must be specified and \code{locusColumn} 
+#' must be correct. Otherwise, \code{distToNearest} will be run with bulk sequencing assumptions, 
+#' using all input sequences regardless of the values in the \code{locusColumn} column.
 #' 
-#' To invoke single-cell mode, both \code{cellIdColumn} and \code{locusColumn} must be supplied. 
-#' Otherwise, the function will run under non-single-cell mode.
+#' Under single-cell mode, only heavy/long chain (IGH, TRB, TRD) sequences will be used for calculating 
+#' nearest neighbor distances. Under non-single-cell mode, all input sequences will be used for 
+#' calculating nearest neighbor distances, regardless of the values in the \code{locusColumn} field (if present).
 #' 
-#' Under single-cell mode, only heavy chain sequences will be used for calculating nearest neighbor
-#' distances. Under non-single-cell mode, all input sequences will be used for calculating nearest
-#' neighbor distances, regardless of the values in the \code{locusColumn} field (if present).
+#' Values in the \code{locusColumn} must be one of \code{c("IGH", "IGI", "IGK", "IGL")} for BCR 
+#' or \code{c("TRA", "TRB", "TRD", "TRG")} for TCR sequences. Otherwise, the function returns an 
+#' error message and stops.
 #' 
 #' For single-cell mode, the input format is the same as that for \link[alakazam]{groupGenes}. 
 #' Namely, each row represents a sequence/chain. Sequences/chains from the same cell are linked
-#' by a cell ID in the \code{cellIdColumn} field. Under this mode, there is a choice of whether 
-#' grouping should be done using heavy chain (\code{IGH}) sequences only, or using both 
-#' heavy chain (\code{IGH}) and light chain (\code{IGK}, \code{IGL}) sequences. This is governed 
-#' by \code{groupUsingOnlyIGH}.
+#' by a cell ID in the \code{cellIdColumn} field. In this mode, there is a choice of whether 
+#' grouping should be done by (a) using IGH (BCR) or TRB/TRD (TCR) sequences only or
+#' (b) using IGH plus IGK/IGL (BCR) or TRB/TRD plus TRA/TRG (TCR). 
+#' This is governed by the \code{onlyHeavy} argument.
 #' 
-#' If used, values in the \code{locusColumn} column must be one of \code{"IGH"}, \code{"IGK"}, and \code{"IGL"}.
+#' Note, \code{distToNearest} required that each cell (each unique value in \code{cellIdColumn})
+#' correspond to only a single \code{IGH} (BCR) or \code{TRB/TRD} (TCR) sequence.
 #' 
-#' Note that for \code{distToNearest}, a cell with multiple heavy chains is not allowed.
-#' 
-#' The distance to nearest (heavy chain) neighbor can be used to estimate a threshold for assigning 
+#' The distance to nearest neighbor can be used to estimate a threshold for assigning 
 #' Ig sequences to clonal groups. A histogram of the resulting vector is often bimodal, with the 
 #' ideal threshold being a value that separates the two modes.
 #' 
@@ -751,21 +759,21 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 #'                                SHazaM v0.1.4 and Change-O v0.3.3.
 #' }
 #' 
-#' Note on \code{NA}s: if, for a given combination of V gene, J gene, and sequence length,
+#' Note on \code{NA}s: if, for a given combination of V gene, J gene, and junction length,
 #' there is only 1  heavy chain sequence (as defined by \code{sequenceColumn}), \code{NA} is 
-#' returned instead of a distance (since it has no heavy chain neighbor). If for a given combination 
-#' there are multiple heavy chain sequences but only 1 unique one, (in which case every heavy cahin 
+#' returned instead of a distance (since it has no heavy/long chain neighbor). If for a given combination 
+#' there are multiple heavy/long chain sequences but only 1 unique one, (in which case every heavy/long cahin 
 #' sequence in this group is the de facto nearest neighbor to each other, thus giving rise to distances 
 #' of 0), \code{NA}s are returned instead of zero-distances.
 #' 
-#' Note on \code{subsample}: Subsampling is performed independently in each VJL group for heavy chain
-#' sequences. If \code{subsample} is larger than number of heavy chain sequences in the group, it is 
-#' ignored. In other words, subsampling is performed only on groups in which the number of heavy chain 
-#' sequences is equal to or greater than \code{subsample}. \code{dist_nearest} has values calculated 
-#' using all heavy chain sequences in the group for groups with fewer than \code{subsample} heavy chain
-#' sequences, and values calculated using a subset of heavy chain sequences for the larger groups. 
-#' To select a value of \code{subsample}, it can be useful to explore the group sizes in \code{db} 
-#' (and the number of heavy chain sequences in those groups).
+#' Note on \code{subsample}: Subsampling is performed independently in each V-J-length group for 
+#' heavy/long chain sequences. If \code{subsample} is larger than number of heavy/long chain sequences 
+#' in the group, it is ignored. In other words, subsampling is performed only on groups in which the 
+#' number of heavy/long chain sequences is equal to or greater than \code{subsample}. \code{dist_nearest} 
+#' has values calculated using all heavy chain sequences in the group for groups with fewer than 
+#' \code{subsample} heavy/long chain sequences, and values calculated using a subset of heavy/long chain 
+#' sequences for the larger groups. To select a value of \code{subsample}, it can be useful to explore 
+#' the group sizes in \code{db} (and the number of heavy/long chain sequences in those groups).
 #' 
 #' @references
 #' \enumerate{
@@ -813,9 +821,8 @@ distToNearest <- function(db, sequenceColumn="junction", vCallColumn="v_call", j
                           normalize=c("len", "none"), symmetry=c("avg", "min"),
                           first=TRUE, VJthenLen=TRUE, nproc=1, fields=NULL, cross=NULL, 
                           mst=FALSE, subsample=NULL, progress=FALSE,
-                          cellIdColumn=NULL, locusColumn=NULL, 
-                          groupUsingOnlyIGH=TRUE, keepVJLgroup=TRUE) {
-    
+                          cellIdColumn=NULL, locusColumn="locus", 
+                          onlyHeavy=TRUE, keepVJLgroup=TRUE) {
     # Hack for visibility of foreach index variables
     i <- NULL
     
@@ -824,43 +831,58 @@ distToNearest <- function(db, sequenceColumn="junction", vCallColumn="v_call", j
     normalize <- match.arg(normalize)
     symmetry <- match.arg(symmetry)
     if (!is.data.frame(db)) { stop('Must submit a data frame') }
-    
-    
-    # single-cell mode?
-    if ( !is.null(cellIdColumn) & !is.null(locusColumn) ) {
-        singleCell <- TRUE
-        
-        if (!all(db[[locusColumn]] %in% c("IGH", "IGK", "IGL"))) {
-            stop("The locus column must be one of {IGH, IGK, IGL}.")
-        }
-        
-    } else {
-        singleCell <- FALSE
-    } 
-    
-    columns <- c(sequenceColumn, vCallColumn, jCallColumn, fields, cross,
-                 cellIdColumn, locusColumn)
-    columns <- columns[!is.null(columns)]
-    
-    check <- checkColumns(db, columns)
+
+    # Check base input
+    check <- checkColumns(db, c(sequenceColumn, vCallColumn, jCallColumn, fields, cross))
     if (check != TRUE) { stop(check) }
     
-    # cast all columns to character
+    # Check single-cell input
+    if (!is.null(cellIdColumn)) {
+        check <- checkColumns(db, c(cellIdColumn, locusColumn))
+        if (check != TRUE) { stop(check) }
+    }
+    
+    # Cast all columns to character
+    columns <- c(sequenceColumn, vCallColumn, jCallColumn, fields, cross,
+                 cellIdColumn, locusColumn)
+    columns <- columns[!is.null(columns) & columns %in% names(db)]
     for (cl in columns) { db[[cl]] <- as.character(db[[cl]]) }
     
     # Convert sequence columns to uppercase
     db <- toupperColumns(db, c(sequenceColumn)) 
     
+    # Single-cell mode?
+    if (!is.null(cellIdColumn) & !is.null(locusColumn)) {
+        singleCell <- TRUE
+        
+        # check locus column
+        valid_loci <- c("IGH", "IGI", "IGK", "IGL", "TRA", "TRB", "TRD", "TRG")
+        check <- !all(unique(db[[locusColumn]]) %in% valid_loci)
+        if (check) {
+            stop("The locus column contains invalid loci annotations.")
+        }
+    } else {
+        singleCell <- FALSE
+    } 
+    
     # Disallow multiple heavy chains per cell
     if (singleCell) {
-        bool <- sapply(unique(db[[cellIdColumn]]), function(x) { 
-            return( sum( db[[locusColumn]][db[[cellIdColumn]]==x] == "IGH" )>1 ) 
-        } )
-        if (any(bool)) {
-            stop("Detected multiple heavy chains in cell(s). Each cell must contain only 1 heavy chain.")
+        # check multiple heavy chains
+        x <- sum(table(db[[cellIdColumn]][db[[locusColumn]] == "IGH"]) > 1)
+        if (x > 0) {
+            stop(paste(x, "cell(s) with multiple heavy chains found. One heavy chain per cell is expected."))
+        }
+        # check multiple beta chains
+        x <- sum(table(db[[cellIdColumn]][db[[locusColumn]] == "TRB"]) > 1)
+        if (x > 0) {
+            stop(paste(x, "cell(s) with multiple beta chains found. One beta chain per cell is expected."))
+        }
+        # check multiple delta chains
+        x <- sum(table(db[[cellIdColumn]][db[[locusColumn]] == "TRD"]) > 1)
+        if (x > 0) {
+            stop(paste(x, "cell(s) with multiple delta chains found. One delta chain per cell is expected."))
         }
     }
-    
     
     # Check for invalid characters
     # heavy
@@ -882,7 +904,7 @@ distToNearest <- function(db, sequenceColumn="junction", vCallColumn="v_call", j
         # V+J only first
         # creates $vj_group
         db <- groupGenes(db, v_call=vCallColumn, j_call=jCallColumn, junc_len=NULL,
-                         cell_id=cellIdColumn, locus=locusColumn, only_igh=groupUsingOnlyIGH,
+                         cell_id=cellIdColumn, locus=locusColumn, only_heavy=onlyHeavy,
                          first=first)
         # L (later)  
         group_cols <- c("vj_group", junc_len)
@@ -892,7 +914,7 @@ distToNearest <- function(db, sequenceColumn="junction", vCallColumn="v_call", j
         # creates $vj_group
         # note that despite the name (VJ), this is based on V+J+L
         db <- groupGenes(db, v_call=vCallColumn, j_call=jCallColumn, junc_len=junc_len,
-                         cell_id=cellIdColumn, locus=locusColumn, only_igh=groupUsingOnlyIGH,
+                         cell_id=cellIdColumn, locus=locusColumn, only_heavy=onlyHeavy,
                          first=first)
         group_cols <- c("vj_group")
     }
@@ -996,7 +1018,9 @@ distToNearest <- function(db, sequenceColumn="junction", vCallColumn="v_call", j
         
         crossGroups <- NULL
         if (!is.null(cross)) {
-            crossGroups <- db_group %>% dplyr::group_indices(!!!rlang::syms(cross))
+            crossGroups <- db_group %>% 
+              dplyr::group_by(!!!rlang::syms(cross)) %>%
+              dplyr::group_indices()
         }
         
         arrSeqs <-  db[[sequenceColumn]][idx]
@@ -1014,7 +1038,7 @@ distToNearest <- function(db, sequenceColumn="junction", vCallColumn="v_call", j
         return(db_group)
     }, 
     error = function(e) {
-      if ( nproc>1 & grepl("Error in unserialize(socklist[[n]]) : error reading from connection", e, fixed=TRUE)) {
+      if (nproc > 1 & grepl("Error in unserialize(socklist[[n]]) : error reading from connection", e, fixed=TRUE)) {
         warning("There is an error running the code in parallel. Try with nproc=1.")
       }
       stop(e)